renaming files, clearing old files
