--- supporting_models/README.md
@@ -1,5 +1,0 @@
-## BTSbot/supporting_models
-
-Scores produced by these models are used at metadata input into BTSbot. 
-
-They come from the following repositories: https://github.com/dmitryduev/acai, https://github.com/dmitryduev/braai
