--- train_configs/mm_maxvit.json
@@ -1,61 +1,0 @@
-{
-  "model_name": "mm_MaxViT",
-  "model_kind": "maxvit_tiny_rw_224.sw_in1k",
-
-  "train_data_version": "v10",
-  "loss": "binary_crossentropy",
-
-  "learning_rate": 1e-6,
-  "warmup_epochs": 4,
-  "beta_1": 0.9,
-  "beta_2": 0.999,
-
-  "batch_size": 16,
-  "epochs": 30,
-  "patience": 6,
-  "N_max": 100,
-  
-  "metadata_cols": [
-    "sgscore1",
-    "distpsnr1",
-    "sgscore2",
-    "distpsnr2",
-    "fwhm",
-    "magpsf",
-    "sigmapsf",
-    "chipsf",
-    "ra",
-    "dec",
-    "diffmaglim",
-    "ndethist",
-    "nmtchps",
-    "age",
-    "days_since_peak",
-    "days_to_peak",
-    "peakmag_so_far",
-    "new_drb",
-    "ncovhist",
-    "nnotdet",
-    "chinr",
-    "sharpnr",
-    "scorr",
-    "sky",
-    "maxmag_so_far"
-  ],
-  
-  "meta_fc1_neurons": 256,
-  "meta_dropout": 0.63,
-  "meta_fc2_neurons": 8,
-  
-  "comb_fc1_neurons": 128,
-  "comb_fc2_neurons": 32,
-  "comb_dropout": 0.6,
-
-  "data_aug_h_flip": 1,
-  "data_aug_v_flip": 1,
-  "data_aug_rot": 1,
-  
-  "random_seed": 2,
-  "generate_embeddings": 1,
-  "testing": 0
-}
