--- train_configs/mm_swin.json
@@ -1,61 +1,0 @@
-{
-  "model_name": "mm_SwinV2",
-  "swin_kind": "swin_v2_t",
-  "swin_weights": "IMAGENET1K_V1",
-
-  "train_data_version": "v11",
-  "loss": "binary_crossentropy",
-  "data_base_dir": "",
-
-  "learning_rate": 5e-5,
-  "beta_1": 0.99,
-  "beta_2": 0.99,
-  "batch_size": 64,
-
-  "epochs": 50,
-  "patience": 6,
-  "N_max_p": 100,
-  "N_max_n": 100,
-  
-  "metadata_cols": [
-    "sgscore1",
-    "distpsnr1",
-    "sgscore2",
-    "distpsnr2",
-    "fwhm",
-    "magpsf",
-    "sigmapsf",
-    "chipsf",
-    "ra",
-    "dec",
-    "diffmaglim",
-    "ndethist",
-    "nmtchps",
-    "age",
-    "days_since_peak",
-    "days_to_peak",
-    "peakmag_so_far",
-    "new_drb",
-    "ncovhist",
-    "nnotdet",
-    "chinr",
-    "sharpnr",
-    "scorr",
-    "sky",
-    "maxmag_so_far"
-  ],
-  
-  "meta_fc1_neurons": 128,
-  "meta_dropout": 0.25,
-  "meta_fc2_neurons": 128,
-  
-  "comb_fc_neurons": 8,
-  "comb_dropout": 0.20,
-
-  "data_aug_h_flip": 1,
-  "data_aug_v_flip": 1,
-  "data_aug_rot": 1,
-  
-  "random_seed": 2,
-  "testing": 0
-}
