--- train_configs/swin.json
@@ -1,29 +1,0 @@
-{
-  "model_name": "SwinV2",
-  "swin_kind": "swin_v2_t",
-  "swin_weights": "IMAGENET1K_V1",
-
-  "train_data_version": "v11",
-  "loss": "binary_crossentropy",
-  "data_base_dir": "",
-  "learning_rate": 1e-4,
-  "beta_1": 0.99,
-  "beta_2": 0.99,
-  "batch_size": 64,
-
-  "epochs": 75,
-  "patience": 15,
-  "N_max_p": 100,
-  "N_max_n": 100,
-  
-  "fc1_neurons": 128,
-  "fc2_neurons": 8,
-  "dropout1": 0.4,
-
-  "data_aug_h_flip": 1,
-  "data_aug_v_flip": 1,
-  "data_aug_rot": 1,
-  
-  "random_seed": 2,
-  "testing": 0
-}
