--- train_configs/um_cnn_config.json
@@ -1,39 +1,0 @@
-{
-    "model_name": "um_cnn",
-    "train_data_version": "v10",
-    "loss": "binary_crossentropy",
-    "learning_rate": 1e-4,
-    "beta_1": 0.999,
-    "beta_2": 0.999,
-    "batch_size": 32,
-
-    "crop_cutout_size": 3,
-
-    "epochs": 1000,
-    "patience": 20,
-    "N_max_p": 100,
-    "N_max_n": 100,
-
-    "conv1_channels": 64,
-    "conv2_channels": 16,
-    "conv_kernel": 3,
-    "conv_dropout1": 0.45,
-    "conv_dropout2": 0.65,
-        
-    "head_neurons": 128,
-    "head_dropout": 0.45,
-
-    "weight_classes": "alert",
-    "reduce_LR_factor": 0.4,
-    "reduce_LR_minLR": 5e-10,
-    "data_aug_h_flip": 1,
-    "data_aug_v_flip": 1,
-    "data_aug_rot": 1,
-
-    "val_sne_only": 0,
-    "val_keep_near_threshold": 1,
-    "val_rise_only": 0,
-    
-    "random_seed": 2
-  }
-
